requirements.txt
asgiref
djangorestframework
Django
django-filter
psycopg2-binary==2.8.5
pytz
sqlparse==0.3.1
gunicorn==20.0.4
typing
celery
redis
django-celery-beat
django_celery_results
flower
names
